+++ python/cudf/cudf/io/__init__.py
@@ -7,4 +7,8 @@ from cudf.io.feather import read_feather
 from cudf.io.hdf import read_hdf
 from cudf.io.json import read_json
 from cudf.io.orc import read_orc, read_orc_metadata, to_orc
-from cudf.io.parquet import read_parquet, read_parquet_metadata
+from cudf.io.parquet import (
+    read_parquet,
+    read_parquet_metadata,
+    write_to_dataset,
+)
